Extract stubDescriptor to own function
