+++ lib/sinon/stub.js
@@ -11,35 +11,19 @@
 var behavior = require("./behavior");
 var spy = require("./spy");
 var extend = require("./util/core/extend");
-var objectKeys = require("./util/core/object-keys");
 var createInstance = require("./util/core/create");
-var deprecated = require("./util/core/deprecated");
 var functionToString = require("./util/core/function-to-string");
 var valueToString = require("./util/core/value-to-string");
 var wrapMethod = require("./util/core/wrap-method");
 var stubEntireObject = require("./stub-entire-object");
+var stubDescriptor = require("./stub-descriptor");
 
 function stub(object, property, descriptor) {
     var isStubbingEntireObject = typeof property === "undefined" && typeof object === "object";
+    var isStubbingDescriptor = Boolean(descriptor);
 
     if (isStubbingEntireObject) {
         return stubEntireObject(stub, object);
-    }
-
-    if (!!descriptor && typeof descriptor === "function") {
-        deprecated.printWarning(
-          "sinon.stub(obj, 'meth', fn) is deprecated and will be removed from" +
-          "the public API in a future version of sinon." +
-          "\n Use stub(obj, 'meth').callsFake(fn)." +
-          "\n Codemod available at https://github.com/hurrymaplelad/sinon-codemod"
-        );
-    }
-    if (!!descriptor && typeof descriptor !== "function" && typeof descriptor !== "object") {
-        throw new TypeError("Custom stub should be a property descriptor");
-    }
-
-    if (typeof descriptor === "object" && objectKeys(descriptor).length === 0) {
-        throw new TypeError("Expected property descriptor to have at least one key");
     }
 
     if (property && !object) {
@@ -51,26 +35,16 @@ function stub(object, property, descriptor) {
         return stub.create();
     }
 
-    var wrapper;
-    if (descriptor) {
-        if (typeof descriptor === "function") {
-            wrapper = spy && spy.create ? spy.create(descriptor) : descriptor;
-        } else {
-            wrapper = descriptor;
-            if (spy && spy.create) {
-                var types = objectKeys(wrapper);
-                for (var i = 0; i < types.length; i++) {
-                    wrapper[types[i]] = spy.create(wrapper[types[i]]);
-                }
-            }
-        }
-    } else {
-        var stubLength = 0;
-        if (typeof object === "object" && typeof object[property] === "function") {
-            stubLength = object[property].length;
-        }
-        wrapper = stub.create(stubLength);
+    if (isStubbingDescriptor) {
+        return stubDescriptor.apply(null, arguments);
     }
+
+    var wrapper;
+    var stubLength = 0;
+    if (typeof object === "object" && typeof object[property] === "function") {
+        stubLength = object[property].length;
+    }
+    wrapper = stub.create(stubLength);
 
     return wrapMethod(object, property, wrapper);
 }
